# api_docs/api-docs-input.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/refresh:
    post:
      tags:
      - login-controller
      summary: Checks if the user logged in is the same as the user in the token
      operationId: refreshToken
      parameters:
      - name: refreshToken
        in: cookie
        required: false
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
          content:
            application/json: {}
        "200":
          description: User is logged in
          content:
            application/json: {}
        "403":
          description: Forbidden
          content:
            application/json: {}
        "400":
          description: Bad request
          content:
            application/json: {}
...
  /api/login:
    post:
      tags:
      - login-controller
      summary: User login through JSON
      operationId: loginForm_1
      parameters:
      - name: accessToken
        in: cookie
        required: false
        schema:
          type: string
      - name: refreshToken
        in: cookie
        required: false
        schema:
          type: string
      - name: loginRequest
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/LoginRequest'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: User logged in successfully
          content:
            application/json: {}
        "401":
          description: Unauthorized
          content:
            application/json: {}
        "403":
          description: Forbidden
          content:
            application/json: {}
        "400":
          description: Bad request
          content:
            application/json: {}
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...